Move controller-specific code to its namespace.
...and out of the trajectory_interface namespace.
